move Go package ./pkg/inventory to underneath ./cmd/frontend/...
Only the frontend uses the inventory (which is the listing of languages used in a repository), so it can and should become an internal package under ./cmd/frontend.
