+++ cmd/frontend/internal/inventory/filelang/LICENSE
