+++ cmd/frontend/internal/inventory/filelang/_data/languages.yml
