+++ cmd/frontend/internal/inventory/filelang/_data/vendor.yml
