+++ cmd/frontend/internal/inventory/filelang/data_languages.go
