+++ cmd/frontend/internal/inventory/filelang/data_vendor.go
