+++ cmd/frontend/internal/inventory/filelang/languages.go
