+++ cmd/frontend/internal/inventory/inventory_test.go
