+++ cmd/frontend/internal/inventory/prom-repo-tree.txt
